refactor: split generate_post.py into three modules (#64)
Extract content assembly into post_content.py (336 lines) and template
rendering into post_templates.py (167 lines), keeping the CLI entry
point in generate_post.py (158 lines — well under the 500-line limit).
